Refactor: error resolve
