+++ .gitignore
@@ -5,6 +5,8 @@ build/
 !**/src/main/**/build/
 !**/src/test/**/build/
 application.properties
+application-dev.properties
+
 ### STS ###
 .apt_generated
 .classpath
