+++ src/main/java/com/sejong/creativesemester/comment/entity/Comment.java
@@ -11,11 +11,10 @@ import lombok.NoArgsConstructor;
 import javax.persistence.*;
 
 @Getter
-@Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Table(name = "COMMENT_TABLE")
+@Entity(name = "COMMENT_TABLE")
 public class Comment extends BaseTimeEntity {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long commentId;
