+++ src/main/java/com/sejong/creativesemester/council/reposiotory/CouncilRepository.java
@@ -1,8 +1,7 @@
-package com.sejong.creativesemester.coucil.reposiotory;
+package com.sejong.creativesemester.council.reposiotory;
 
-import com.sejong.creativesemester.coucil.entity.Council;
+import com.sejong.creativesemester.council.entity.Council;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
