+++ src/main/java/com/sejong/creativesemester/council/reposiotory/CouncilRepositoryCustom.java
@@ -1,6 +1,6 @@
-package com.sejong.creativesemester.coucil.reposiotory;
+package com.sejong.creativesemester.council.reposiotory;
 
-import com.sejong.creativesemester.coucil.reposiotory.res.CouncilInfoResponseDto;
+import com.sejong.creativesemester.council.reposiotory.res.CouncilInfoResponseDto;
 
 public interface CouncilRepositoryCustom {
     CouncilInfoResponseDto findCouncilInfoByMajorId(Long majorId);
