+++ src/main/java/com/sejong/creativesemester/major/entity/Major.java
@@ -10,11 +10,10 @@ import javax.persistence.*;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Major {
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @Column(nullable = false)
-    private String majorName;
+    private String name;
 }
